refactors analytics functionality into core
