+++ src/matcha_ml/cli/cli.py
@@ -4,7 +4,6 @@ from typing import Optional
 import typer
 
 from matcha_ml import __version__
-from matcha_ml.cli import analytics
 from matcha_ml.cli._validation import (
     prefix_typer_callback,
     region_typer_callback,
@@ -24,10 +23,9 @@ from matcha_ml.core import core
 from matcha_ml.errors import MatchaError, MatchaInputError
 
 app = typer.Typer(no_args_is_help=True, pretty_exceptions_show_locals=False)
-
-# Create a group for all subcommands for analytics command
+analytics_app = typer.Typer(no_args_is_help=True, pretty_exceptions_show_locals=False)
 app.add_typer(
-    analytics.app,
+    analytics_app,
     name="analytics",
     help="Enable or disable the collection of anonymous usage data (enabled by default).",
 )
@@ -131,6 +129,24 @@ def cli(
     For more help on how to use matcha, head to https://fuzzylabs.github.io/matcha/
     """
     pass
+
+
+@analytics_app.command()
+def opt_out() -> None:
+    """Disable the collection of anonymous usage data."""
+    print(
+        "Data collection has been turned off and no data will be collected - you can turn this back on by running the command: 'matcha analytics opt-in'."
+    )
+    core.analytics_opt_out()
+
+
+@analytics_app.command()
+def opt_in() -> None:
+    """Enable the collection of anonymous usage data (enabled by default)."""
+    print(
+        "Thank you for enabling data collection, this helps us improve matcha and anonymously understand how people are using the tool."
+    )
+    core.analytics_opt_in()
 
 
 if __name__ == "__main__":
