+++ tests/test_core/test_core.py
@@ -1,11 +1,18 @@
 """Test suite to test the core matcha commands and all its subcommands."""
 import json
 import os
+from typing import Dict, Union
+from unittest import mock
 
 import pytest
+import yaml
 
+from matcha_ml.cli.cli import app
 from matcha_ml.core.core import get
 from matcha_ml.errors import MatchaError, MatchaInputError
+from matcha_ml.services.global_parameters_service import GlobalParameters
+
+INTERNAL_FUNCTION_STUB = "matcha_ml.services.global_parameters_service.GlobalParameters"
 
 
 @pytest.fixture(autouse=True)
@@ -54,6 +61,22 @@ def expected_outputs() -> dict:
     return outputs
 
 
+@pytest.fixture(autouse=True)
+def teardown_singleton():
+    """Tears down the singleton before each test case by clearing the current object."""
+    GlobalParameters._instance = None
+
+
+@pytest.fixture
+def expected_configuration() -> Dict[str, Union[str, bool]]:
+    """Pytest fixture to return expected configuration.
+
+    Returns:
+        Dict[str, Union[str, bool]]: Dictionary containing expected configuration
+    """
+    return {"analytics_opt_out": False, "user_id": "dummy_user_id"}
+
+
 def test_get_resources(expected_outputs: dict):
     """Test get resources function with no resource specified.
 
@@ -92,3 +115,87 @@ def test_get_resources_with_resource_name_and_property_name():
     expected_output = {"experiment-tracker": {"tracking-url": "mlflow_test_url"}}
 
     assert expected_output == get("experiment-tracker", "tracking-url")
+
+
+def test_opt_out_subcommand(
+    runner,
+    matcha_testing_directory: str,
+    expected_configuration: Dict[str, Union[str, bool]],
+) -> None:
+    """Test opt-out command works.
+
+    Args:
+        runner: Mock runner
+        matcha_testing_directory (str): Temp directory
+        expected_configuration (Dict[str, Union[str, bool]]): Dictionary containing expected configuration
+    """
+    config_file_path = os.path.join(
+        matcha_testing_directory, ".matcha-ml", "config.yaml"
+    )
+
+    expected_configuration["analytics_opt_out"] = True
+
+    # Check if config file is not present
+    assert not os.path.exists(config_file_path)
+
+    with mock.patch(
+        f"{INTERNAL_FUNCTION_STUB}.default_config_file_path",
+        new_callable=mock.PropertyMock,
+    ) as file_path, mock.patch(
+        f"{INTERNAL_FUNCTION_STUB}.user_id",
+        new_callable=mock.PropertyMock,
+    ) as uid:
+        file_path.return_value = config_file_path
+        uid.return_value = "dummy_user_id"
+        result = runner.invoke(app, ["analytics", "opt-out"])
+
+        # Check if running command is success
+        assert result.exit_code == 0
+
+    # Check if config file is present
+    assert os.path.exists(config_file_path)
+
+    # Check the contents of the config file match
+    with open(config_file_path) as f:
+        assert dict(yaml.safe_load(f)) == expected_configuration
+
+
+def test_opt_in_subcommand(
+    runner,
+    matcha_testing_directory: str,
+    expected_configuration: Dict[str, Union[str, bool]],
+) -> None:
+    """Test opt-in command works.
+
+    Args:
+        runner: Mock runner
+        matcha_testing_directory (str): Temp directory
+        expected_configuration (Dict[str, Union[str, bool]]): Dictionary containing expected configuration
+    """
+    config_file_path = os.path.join(
+        matcha_testing_directory, ".matcha-ml", "config.yaml"
+    )
+
+    # Check if config file is not present
+    assert not os.path.exists(config_file_path)
+
+    with mock.patch(
+        f"{INTERNAL_FUNCTION_STUB}.default_config_file_path",
+        new_callable=mock.PropertyMock,
+    ) as file_path, mock.patch(
+        f"{INTERNAL_FUNCTION_STUB}.user_id",
+        new_callable=mock.PropertyMock,
+    ) as uid:
+        file_path.return_value = config_file_path
+        uid.return_value = "dummy_user_id"
+        result = runner.invoke(app, ["analytics", "opt-in"])
+
+        # Check if running command is success
+        assert result.exit_code == 0
+
+    # Check if config file is present
+    assert os.path.exists(config_file_path)
+
+    # Check the contents of the config file match
+    with open(config_file_path) as f:
+        assert dict(yaml.safe_load(f)) == expected_configuration
